+++ examples/adapterExample/src/main/java/io/realm/examples/realmadapters/MyAdapter.java
@@ -22,6 +22,7 @@ import android.view.ViewGroup;
 import android.widget.ListAdapter;
 import android.widget.TextView;
 
+import io.realm.OrderedRealmCollection;
 import io.realm.RealmBaseAdapter;
 import io.realm.RealmResults;
 import io.realm.examples.realmadapters.models.TimeStamp;
@@ -32,7 +33,7 @@ public class MyAdapter extends RealmBaseAdapter<TimeStamp> implements ListAdapte
         TextView timestamp;
     }
 
-    public MyAdapter(Context context, int resId, RealmResults<TimeStamp> realmResults, boolean automaticUpdate) {
+    public MyAdapter(Context context, int resId, OrderedRealmCollection<TimeStamp> realmResults, boolean automaticUpdate) {
         super(context, realmResults, automaticUpdate);
     }
 
@@ -48,12 +49,12 @@ public class MyAdapter extends RealmBaseAdapter<TimeStamp> implements ListAdapte
             viewHolder = (ViewHolder) convertView.getTag();
         }
 
-        TimeStamp item = realmResults.get(position);
+        TimeStamp item = adapterData.get(position);
         viewHolder.timestamp.setText(item.getTimeStamp());
         return convertView;
     }
 
-    public RealmResults<TimeStamp> getRealmResults() {
-        return realmResults;
+    public OrderedRealmCollection<TimeStamp> getAdapterData() {
+        return adapterData;
     }
 }